- Separating the helper methods to a separate helper file.
